+++ Cmdr/BuiltInCommands/Utility/history.lua
@@ -2,9 +2,9 @@ return {
 	Name = "history";
 	Aliases = {};
 	AutoExec = {
-		"alias ! run ${history $1}";
-		"alias ^ run ${run replace ${history -1} $1 $2}";
-		"alias !! ! -1";
+		"alias \"!|Displays previous command from history.\" run ${history $1{number|Line Number}}";
+		"alias \"^|Runs the previous command, replacing all occurrences of A with B.\" run ${run replace ${history -1} $1{string|A} $2{string|B}}";
+		"alias \"!!|Reruns the last command.\" ! -1";
 	};
 	Description = "Displays previous commands from history.";
 	Group = "DefaultUtil";